// src/app/model/Scholar.ts
export class Scholar {
    userId: number;
    date: Date;
    name: string;
    lastLogin: string;
    online: boolean;
    image: { darkImage: string; lightImage: string ;darkImageErr: string; lightImageErr: string};
    email:  string;
    role:  string;
    location:  string;
    nationality:  string;
    details: string;
    language:  string;
    specialisationIn:  string;
    studiesAt:  string;
    userEmail:string;
    showHistory:boolean;
    UserLoginStatus:string;
    ImageID:string;  
    Duration:string;
}